Update my AUCTeX setup, preview-latex, etc.
Moved from using Evince -> PDF Tools.
Add "--synctex=1" to be able to navigate back and forth between the
two.

Add some code to better scale preview-latex fragments. For some reason
it wasn't detecting font scale properly.

Add a note about preview-latex fragments not using the correct font
color, especially with a light-on-dark color theme. It turns out the
issue was that ghostscript needs to be > 9.27.